+++ docs/source/News.rst
@@ -24,3 +24,5 @@ This document describes the releases of :mod:`repoze.what`.
     :class:`repoze.what.predicates.NotAuthorizedError` (it replaces
     :class:`repoze.what.predicates.PredicateError`).
 * Introduced :meth:`repoze.what.predicates.Predicate.is_met`.
+* Now :meth:`repoze.what.predicates.Predicate.unmet` receives an optional
+  argument to override the error message.